+++ earth2studio/models/dx/__init__.py
@@ -28,6 +28,7 @@ from earth2studio.models.dx.derived import (
 from earth2studio.models.dx.identity import Identity  # noqa
 from earth2studio.models.dx.precipitation_afno import PrecipitationAFNO  # noqa
 from earth2studio.models.dx.precipitation_afno_v2 import PrecipitationAFNOv2  # noqa
+from earth2studio.models.dx.solarradiation_afno import SolarRadiationAFNO  # noqa
 from earth2studio.models.dx.tc_tracking import (
     TCTrackerVitart,
     TCTrackerWuDuan,